Moving vwilight theme to 'vwilight-original' and removing changes
There is a newer version of this theme that will be used. In order
to be able to use both at the same time, moving this into a different
location so it can be used with a different name.

Also removing customizations so there is a clearer record of the
file's history.