Moving all home-related dotfiles to `files` directory
Because there are housekeeping files in the root level of the
repository, particularly git files with the name name as git config
dot files, moving all of the dot files into a `files` directory.

This `files` directory has the same structure as the files are intended
to be placed in the home directory, which makes it easy to symlink them
without needing to determine which files should be symlinked or not.

This also makes it possible to keep the original filenames, which should
help avoid confusion.

In order to make the symlink and navigation structure slightly cleaner,
this repo can live at `~/.dot`, so then all of the files at the
`~/.dot/files/` path.