Disabling sudo in in scripts/brew.sh
Homebrew is intended to be used without sudo, so it's not entirely clear
why sudo is being called throughout this script. Currently keeping all
of the original commands in the file, though commented out, in order to
ensure that this approach works before removing all of them entirely. In
one case, a command was called without sudo.