Creating separate osx_app.sh script for non-default app prefs
Also adding setting for secret iTerm2 preference for alternate
scrolling.

Splitting osx.sh and osx_apps.sh provides a better distinction between
preferences for a fresh default install of OS X and preferences for
applications that need to be installed separately. As part of this
process, moving the aliasing of iOS Simulator to the new osx_app.sh
script since it requires Xcode to be installed on fresh copies of OS X.

Also adding a new task in the manuelfile