Adding .local.examples; gitignoring known .local files.
In order to more explicitly handle .local dotfiles, adding examples that
can be copied and used for local settings.

The downside to this approach is that the home directories end up with a
few unused "example" files, but this tradeoff seems reasonable for the
benefit of explitness.

Also adding known local files to .gitignore in case they inadvertently
end up getting added to the repository.