Reconfigure golang install for homebrew
Previous to this, golang was installed using the OS X package. However,
that approach is out of sync with how most packages in this system are
installed and configured, which leads to some dependency issues if
versions are not properly in sync. As a result, bringing golang under
the homebrew umbrella.

Additionally, in order to separate go projects from whatever particular
install method is being used at any time, $GOPATH and everything related
to it are set to $HOME/.go with the intention that external symlinks
make a connection from external directories that need to be aware of
where something in $GOPATH is located.

Finally, adding a go-packages install script with the `vet` and `godoc`
commands, as noted in the homebrew formula caveats: http://git.io/sxunsw