Add .config/htop/htoprc as link to .htoprc
~/.config is apparently the only recognized location for htoprc on
ubuntu, so creating a symlink there to point to ~/.htoprc appears to be
the only way to get ubuntu htop to recognize the config file.

Since ~/.config contains many config files that may or may not be ideal
to track, everything else in this directory should be gitignored at this
stage.