Add autoconf to brew install list
This is currently needed to build old versions of Ruby. It was already
installed on the system, but appears to have been only partially linked.
The reason for this wasn't investigated since relinking it resolved the
issue. Adding to the list explicitly should indicate that this is
expected to be on the system, and should help resolve issues on other
machines with this configuration.