Add cabal config as templates
The templates included in this commit can be used to generate the actual
cabal configuration files. Templates are useful here because the cabal
configuration files expect to have full, absolute paths, which ends up
making them hard-coded to a specific user account. By using a simple
script to generate the files from templates, this issues around these
being user-specific can be avoided.

The script itself is included as part of the install/cabal-packages
script so it ends up being run every time this package script is run.