Re-add right scrollbar in MacVim
When the right scrollbar is disabled, part of the scrollbar background
is still displayed in white when MacVim is in full screen mode. This is
distracting and also has not obvious functionality. Although the full
scrollbar is also distracting and white/bright, it does provide some
funtionality to make up for it. Therefore, including the scrollbar makes
more sense than removing it until disabling it completely removes it in
all contexts.