Add Sublime Text preferences comments as JSON
Even though JSON doesn't support comments, Sublime Text supposedly
allows them in the preferences files. However, the added comments appear
to be getting removed, either by Sublime Text or a linter. In order to
avoid this issue, comments can be added as key / value pairs in the JSON
itself. The convention I'm using is to add a new key with the same name
as the key that the comment applies to, but with a `_comment` suffix
appended. With this convention, comments remain adjacent to the key
their apply to when the settings are sorted.