Add opam-packages install script
The initial target package in Irmin, but this also requires installing
the cohttp and git packages.

Also, the presence of `export LDFLAGS='-arch x86_64'` in the environment
results in build failures, so unset this in the opam-packages install
script and include a not in the osx_build_variables.sh file in case this
issue comes up in other contexts.