Move .emacs.d config to vendor and symlink in home
.emacs.d is a submodule for the custom prelude configuration, but in
order to swap the configuration easily, this submodule is better to have
in vendor and use a symlink in the home directory.

Currently the name is wrong because the `git-submodule-move` script was
used placed it in the current location with the '.emacs.d' name. A
renaming should be attempted in a subsequent commit in order to make it
match the origin repository URL.