Set vim is_posix variable to avoid pre-POSIX sh error highlighting
Setting this variable prevents errors from being displayed in sh shell
scripts. For example, without this setting, $() command substitution is
displayed as an error, typically with a red background, because pre-POSIX
shells only supported the backtick syntax (`..`) which is now discouraged.
(see: https://github.com/koalaman/shellcheck/wiki/SC2006). With this
setting, however, the default is set to KornShell (ksh), which is
POSIX-compliant and supports the $() syntax.

 More information:

https://github.com/b4winckler/vim/blob/master/runtime/doc/syntax.txt#L2802
http://stackoverflow.com/a/5423808