Set seti.vim submodule to master branch latest
In a previous commit this was set to the latest master commit, which is
not what is not the preferred commit. This behavior appears to be due to
the way that git syncs submodules, where it defaults to the master
branch and will reset the commit to the latest master commit if it finds
the submodule is currently at a different commit.

The master branch is now set to the same as signify-colors-dark latest,
so future updates should pull from master as expected while retaining
this commit reference as long as it's the latest on master.