Add vendored bashmenot, linked to environment as .dir in $HOME
bashmenot is a collection of bash functions. Rather than automatically
load it into the environment, linking the vendored directory as a dot
directory makes it available and easy to source. Putting the directory
in .bashrc.d or .profile.d would also be suboptimal since the structure
of the project differs from the structure of these directories, and
because every file at the root level of these directories are
intended to be sourced together.

In a subsequent revision it might make sense to create a different
wrapper for these functions, such as a way to source these on demand as
a whole or individually, or possibly make them available through a
separate script.