Upgrade Lua and LuaRocks installations
The homebrew formula for lua now includes LuaRocks, which also uses an
executable with a version number appended to the name, so the
installation of both the homebrew formulae and the rocks in the luarocks
install script need updates to account for these changes.

Since the luarocks fomula has been removed from homebrew but not from
the system, the existing installtion needs to removal.

In order to identify the existence of a formula installation, a new
function is added to use `brew list` to check.