Add `search` script replacing the web-search oh-my-zsh plugin
This introduces a bask script that provides the same functionality as
the web-search oh-my-zsh plugin, but a separate bash executable rather
than a series of commands in the current environment.

This script opens search engines in the terminal using the specified
default terminal browser or, if none exists, falls back to opening an
external browser similar to how the oh-my-zsh plugin works.