Place pathogen plugin disabling before janus and disable gitgutter
The `janus#disable_plugin` function appears to use the
`g:pathogen_disabled` variable, so setting this to an empty value after
the janus calls appears to revert the plugin disabling for janus. This
issue became apparent when janus added the vim-gitgutter plugin since
the pause on focus that it added is apparent and makes it clear that the
plugin is not disabled when the pathogen settings are placed after the
janus ones.

Additionally, in order to disabled the vim-gitgutter plugin the enitre
name with the 'vim-' prefix must be passed to `janus#disable_plugin`.