Remove gpgtools from brew cask install list
gpgtools conflicts with the gnupg2 formula and has a somewhat opaque
installation process that asks for admin privilages. gnupg2 is preferred
since it's the standard way to install this.

Because the gpgtools cask is no longer used, the gnupg2 link overwrite
is no longer needed, either, and is also removed.