Remove .zshenv.d/mactex.zsh
This file runs OS X's `path_helper` program, which prints the default
`$PATH` prepended to the environment's existing `$PATH`. However, it
appears to cause issues due to the command getting called in a subshell,
and `$PATH` ends up having the initial `$PATH` prepended to the desired
`$PATH`. Furthermore, since the mactex path is added to `/etc/paths.d`,
it is apparently automatically loaded with new logins, anyway.
Therefore, it appears this file isn't necessary.