Restructure zsh and bash initialization.
Configuration that would normally be found in either .[z]profile or
.[bash|zsh]rc have been placed in .profile and .profile.d together, with
.profile being sources by both .bashrc and .zshrc. However, .profile
has been sourced in .bashrc and .zshrc since zsh doesn't automatically
source it and bash doesn't source it due to the presence of
.bash_profile.

However, .zprofile is a legitimate zsh startup file that is loaded
before .zshrc. For completeness and better control over the startup
sequence, .zprofile is included and sources everything in .profile.d as
expected.

With .profile.d loaded earlier in the startup sequence, some items that
depend on being later in the rc startup sequence end up with errors. In
order to accommodate them, they are moved to a new .shared_rc.d
directory and sourced in the respective rc files from there.

In order to make it easier to visualize the startup sequence, a
`__verbose_init` function is added in the .shared_verbose_init file.