Move contents of .profile.d to shared_rc.d
Before the addition of .shared_rc.d, .profile was acting as a sort of
shared rc startup file and directory. However, now .profile is used by
zsh as part of the .zprofile step in the startup sequence, and .zprofile
is only intended for use with login shells.rather than all interactive
shells. The commands contained in .profile.d are actually intended for
all interactive shells, so the correct place for them is in shared_rc.d.