Add syntastic.override submodule.
Janus plugins can't be disabled and overridden by a plugin with the same
name. Disabling the janus plugin also disables manually-installed
plugins in .janus. If the janus plugin is not disabled, then the plugin
is loaded twice, with the plugin in the .janus folder being loaded after
the one in the janus distribution.

According to the customization wiki, plugins can override the janus
distribution plugins by being added to .janus.before, but so far I've
been unable to get this to work. For more information see:

https://github.com/carlhuda/janus/wiki/Customization#adding-a-new-plugin-color-scheme-

In order to override it within the .janus folder, rename the syntastic
submodule to syntastic.overrride and retain the disabling of the
traditionally-named syntastic in the janus distribution.