Add `$HOME/.hyperlocal` hierarchy configuration.
The `$HOME/.hyperlocal` hierarchy mirrors the `/usr/local` hierarchy, but
only applied to the current user. Installing programs to this location
avoids polluting `$HOME`. `$HOME/.local` would have been a preferred
name, but this directory is used by GNOME and Ubuntu for application
data, so using this separate directory avoids unexpected conflicts.

The nested `bin` and `man` directories are added to the `$PATH` and
`$MANPATH`, and everything except for the Readme is gitignored.