Add `bash` shebang to `.shared_env`
Although this file is not intended to be run as a script, including a
shebang doesn't adversely affect is and can tell all editors what file
type to use for syntax highlighting and other functionality.

Although `sh` would be a logical file type choice, the bash-specific
features used here are compatible with zsh and provide enough benefit to
keep them around for now.