Use relative path to `bindle` in `bootstrap/linux` `brew-linux` command.
At the point in the bootstrap sequence that this command is encountered,
`bindle` is not yet in the `$PATH` since the `$PATH` is only modified by
the linking in the previous steps. Therefore, an explicit path needs to
be used, just like the other `bindle` calls earlier in the sequence.