Configure z.sh as a submodule rather than as a plugin.
The z.sh plugin for oh-my-zsh doesn't make it easy to control the
initialization cycle of the script, and it makes it more difficult to
determine where it is being loaded from. Including as a submodule makes
it a little more clear what's going on.

The biggest problem with this approach is the that the only place to
include the links to the vendored script and man page is
`~/.hyperlocal`, which is .gitignored. In order to include these links
in this repository, cascading .gitignore statements are required in
order to selectively include the files, and that's less than ideal.