Remove z.sh in favor of fasd.
fasd is already installed and provides similar functionality with an
default alias on `z`. z.sh also currently has a bug where it throws an
error when a file is not found at `~/.z`, which happens in new user
accounts the first time the configuration is loaded. Finally, removing
z.sh simplifies the structure of this configuration and eliminates a
submodule.