Move per-directory-history to vendor and load in .zshrc.d
Installing this as a submodule in .oh-my-zsh.custom is not easily
discoverable and loading with oh-my-zsh seems unnecessary when it can
just be sourced directly during the shell loading process. The
oh-my-zsh approach also resulted in the following error:

    command not found: add-zsh-hook

Sourcing directly avoids this error.

Also, the plugin name was included in the oh-my-zsh plugins list with
underscores, when the correct name uses hyphens.