Remove homebrew completions from ZSH $fpath on user mismatch.
/usr/local/share/zsh/site-functions contains homebrew-installed
completions, but because are are installed with a non-root user account,
any user other than the user who originally installed the formulas will
encounter warnings from `compaudit` and `compinit`. There doesn't appear
to be a good way to handle this since the non-root homebrew philosophy
conflicts with the expectations of ZSH. As a result, if this directory
is present and owned by a different user, then remove it from `$fpath`
before initializing oh-my-zsh.