Add `disk` script to ./bin.
`disk` is currently very similar to the `usage` script, but it provides
subcommands that are focused on specific operations using the proxied
commands with certain options.

One objective behind the addition of this script is to determine whether
there are other operations related to disks, beyond space usage
statistics, than can be included in order to extend the functionality of
this script.