Handle both GNU and BSD versions of `du` in `disk`.
The depth option in GNU `du` is `--max-depth`, but in BSD `du` it's
`-d`. Account for both possible options by testing whether
`du --version` exists, since the `--version` flag only exists on the BSD
side. The `--max-depth` argument can't be tested in this way both
because it requires an argument and causes `du` to run as if called
directly. In contrast, the `--version` option outputs simple version
information and is therefore a lightweight test.