Add `focus` bask script.
In this configuration, the focus-follows-mouse feature is implemented
via Amethyst. As far as I can tell so far, there is no obvious way to
enable or disable the feature while Amethyst is running. Since Amethyst
is really only used for this one feature at this time, fall back to
opening and closing the application as a means to control this behavior.