Replace `src_init` with `source_dir`.
`src_init` was implemented in a confusing way, expecting a path relative
to `$HOME` as the first argument and an optional single file as the
second argument. `source_dir` is much simpler, taking a path to a directory
as the only argument and iterating through its contents sourcing each file.