Update `vim` wrapper and nvim configuration for neovim 0.1.0.
The configuration for neovim has been moved to ~/.config/nvim:

https://neovim.io/doc/user/nvim_from_vim.html

In the current configuration this is not configured because vim uses the
janus distribution, which includes the ~/.janus directory as well as the
~/.vim.before and ~/.vim.after files, for which there is no immediately
obvious way to include in ~/.config/nvim. To avoid running nvim when
the configuration is not available at this location, check for it in the
`vim` wrapper and fall back to `vim` when it's missing.