Add symlinks for neovim configuration.
Neovim's configuration is contained in the home/.config directory, but
consists of symlinks to existing vim configuration files. For more
information, see:

https://neovim.io/doc/user/nvim_from_vim.html

This configuration prefers symlinks with relative paths, but the nested
symlinks with different relative positions is somewhat confusing.
However, the arrangement in this commit appears to work as intended.