Reorder and better document .shared_env "Init" section.
`$HOME/bin` needs to be added to `$PATH` in order for `bindle` to be
available before assignment of the `$DOTFILES` variable, since the
assignment of `$DOTFILES` is done with `bindle dir` when possible.

Expand comment documentation to make sections easier to scan visually.

Add `shellcheck source` directive for ShellCheck SC1090

https://github.com/koalaman/shellcheck/wiki/SC1090