Duplicate prepending of `$HOME/bin` to `$PATH` in home/.shared_env.
`$HOME/bin` must be prepended to `$PATH` at beginning of shell
initialization so its executables are available for use throughout the
process. However, thoughout this process other paths are prepended,
placing `$HOME/bin` further down the chain. `$HOME/bin` is intended to
be at the beginning of `$PATH` so the executables it contains have the
highest precedence. Therefore, it must be re-prepended after the
shared environment initialization process has been completed.