Remove `mdv`.
`mdv` is currently in a pre-alpha state and does not work consistently
with all documents. More importantly, the symlinked executable exits
with an error because it looks for the 'ansi_tables.json' file using a
relative path lookup. Therefore, remove the whole submodule until there
is an opportunity to try to fix some of the issues with `mdv` in this
configuration.