Add mutt configuration.
This configuration is structured for multiple gmail accounts, with
account settings specified by assigning environment variables in a shell
script, as demonstrated in ~/.mutt/accounts/example@example.com.

Gmail-specific settings are split out into a 'provider' file with the
intention of creating a similar provider muttrc file for exchange in the
future.