Replace `mutt` command in shell to avoid accidental direct launch.
The current mutt configuration requires environment variables to be set,
ideally using the `email` shell function. In order to avoid launching
mutt without these being set, set the path to the mutt executable in a
`$MUTT_COMMAND` shell variable and add a `mutt` shell function
instructing the user to use the `email` function.